requirements.txt
dj-database-url==0.5.0
Django==2.1.12
django-cors-headers==3.1.0
django-s3-storage==0.12.5
django-webpack-loader==0.6.0
djangorestframework==3.10.3
flake8==3.7.8
psycopg2==2.8.3
python-decouple==3.1
jsonschema==3.0.2
cryptography==2.7
